requirements.txt
line-bot-sdk
flask
twstock